+++ krkn_ai/chaos_engines/krkn_runner.py
@@ -4,15 +4,15 @@ import datetime
 import tempfile
 
 from krkn_lib.prometheus.krkn_prometheus import KrknPrometheus
-from chaos_ai.chaos_engines.health_check_watcher import HealthCheckWatcher
-from chaos_ai.models.app import CommandRunResult, FitnessResult, FitnessScoreResult, KrknRunnerType
-from chaos_ai.models.config import ConfigFile, FitnessFunctionType
-from chaos_ai.models.scenario.base import Scenario, BaseScenario, CompositeDependency, CompositeScenario
-from chaos_ai.models.scenario.factory import ScenarioFactory
-from chaos_ai.utils import run_shell
-from chaos_ai.utils.fs import env_is_truthy
-from chaos_ai.utils.logger import get_module_logger
-from chaos_ai.utils.rng import rng
+from krkn_ai.chaos_engines.health_check_watcher import HealthCheckWatcher
+from krkn_ai.models.app import CommandRunResult, FitnessResult, FitnessScoreResult, KrknRunnerType
+from krkn_ai.models.config import ConfigFile, FitnessFunctionType
+from krkn_ai.models.scenario.base import Scenario, BaseScenario, CompositeDependency, CompositeScenario
+from krkn_ai.models.scenario.factory import ScenarioFactory
+from krkn_ai.utils import run_shell
+from krkn_ai.utils.fs import env_is_truthy
+from krkn_ai.utils.logger import get_module_logger
+from krkn_ai.utils.rng import rng
 
 logger = get_module_logger(__name__)
