+++ krkn_ai/constants.py
