+++ krkn_ai/models/config.py
@@ -2,9 +2,9 @@ import datetime
 from enum import Enum
 from typing import Dict, List, Optional, Union
 from pydantic import BaseModel, Field, field_validator, model_validator
-import chaos_ai.constants as const
-from chaos_ai.models.cluster_components import ClusterComponents
-from chaos_ai.utils import id_generator
+import krkn_ai.constants as const
+from krkn_ai.models.cluster_components import ClusterComponents
+from krkn_ai.utils import id_generator
 
 
 class PodScenarioConfig(BaseModel):
