+++ krkn_ai/models/custom_errors.py
