+++ krkn_ai/models/scenario/scenario_container.py
@@ -1,6 +1,6 @@
-from chaos_ai.utils.rng import rng
-from chaos_ai.models.scenario.base import Scenario
-from chaos_ai.models.scenario.parameters import *
+from krkn_ai.utils.rng import rng
+from krkn_ai.models.scenario.base import Scenario
+from krkn_ai.models.scenario.parameters import *
 
 
 class ContainerScenario(Scenario):
