+++ krkn_ai/models/scenario/scenario_dummy.py
@@ -1,5 +1,5 @@
-from chaos_ai.models.scenario.base import Scenario
-from chaos_ai.models.scenario.parameters import *
+from krkn_ai.models.scenario.base import Scenario
+from krkn_ai.models.scenario.parameters import *
 
 
 class DummyScenario(Scenario):
