+++ krkn_ai/reporter/health_check_reporter.py
@@ -10,8 +10,8 @@ from matplotlib.ticker import MaxNLocator
 
 from typing import List
 
-from chaos_ai.models.app import CommandRunResult
-from chaos_ai.utils.logger import get_module_logger
+from krkn_ai.models.app import CommandRunResult
+from krkn_ai.utils.logger import get_module_logger
 
 logger = get_module_logger(__name__)
