+++ krkn_ai/templates/krkn-ai.yaml.j2
