+++ krkn_ai/utils/cluster_manager.py
@@ -4,8 +4,8 @@ import ssl
 from typing import List
 from krkn_lib.k8s.krkn_kubernetes import KrknKubernetes
 from kubernetes.client.models import V1PodSpec
-from chaos_ai.utils.logger import get_module_logger
-from chaos_ai.models.cluster_components import ClusterComponents, Container, Namespace, Node, Pod
+from krkn_ai.utils.logger import get_module_logger
+from krkn_ai.models.cluster_components import ClusterComponents, Container, Namespace, Node, Pod
 
 logger = get_module_logger(__name__)
