+++ krkn_ai/utils/rng.py
